refactor: split validation into stages
